Improved the QueryResolver parameter resolution and database resolution functions.
Instead of checking whether each parameter is null, we assign it a specific integer value, 1, 2, or 4. The combination of these values tells us which parameters our resolveQuery() function encountered. This way our queryDatabase() function can know which JPA repository method to call, and with what parameters, based on that value. If 1: only deviceId, if 2, only status, if 3: device and status, and so forth.